refactor: move assessTask to @polpo-ai/core with FileSystem/Shell abstractions
- assessor.ts: replaced exec/access/mkdir/writeFile with Shell.execute()
  and FileSystem methods via AssessmentDeps interface
- scoring.ts, schemas.ts, retry.ts: moved as pure logic (zero Node.js deps)
- llm-review.ts stays in root (pi-ai coupled), injected via runLLMReview port
- Root shims delegate to core with NodeFileSystem/NodeShell defaults
- Cloud can pass SandboxProxyFS/SandboxProxyShell for same assessment in sandbox

All 1153 tests pass, zero typecheck errors.
